chore: update benchmark
