+++ bench/bundle-test/package.json
@@ -1,0 +1,6 @@
+{
+  "private": true,
+  "scripts": {
+    "bench:prepare": "rollup -c && du -h dist/*"
+  }
+}
