--- bundle-test/package.json
@@ -1,6 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "run": "rollup -c && du -h dist/*"
-  }
-}
